requirements-dev.txt
pytest==8.0.0

mypy==1.8.0

black==24.1.1

pytest-dotenv==0.5.2

pytest-asyncio==0.23.5

pytest-cov==4.1.0

pylint==3.0.3